backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.2.0
python-socketio[asgi]>=5.12.0
paho-mqtt==2.1.0
pandas==2.2.3
openpyxl==3.1.5
deepface==0.0.93
tensorflow>=2.16.0
tf-keras
opencv-python==4.8.1.78
scipy==1.11.4
python-multipart==0.0.6
